requirements.txt
django
djangorestframework
channels
daphne
python-binance
python-decouple
pandas
django-cors-headers
